+++ packages/react/src/chains/evm/unichain.ts
@@ -1,0 +1,28 @@
+import { EVMChain } from '../../types/index.js';
+
+export const uniChain: EVMChain = {
+  id: 130,
+  name: 'Unichain',
+  type: 'evm',
+  nativeCurrency: {
+    decimals: 18,
+    name: 'Ether',
+    symbol: 'ETH',
+  },
+  rpcUrls: {
+    default: {
+      http: ['http://mainnet.unichain.org'],
+    },
+  },
+  blockExplorers: {
+    default: {
+      name: 'Unichain Explorer',
+      url: 'https://unichain.blockscout.com',
+    },
+  },
+  contracts: {
+    multicall3: {
+      address: '0xcA11bde05977b3631167028862bE2a173976CA11',
+    },
+  },
+};